backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
psycopg[binary]>=3.2.3
sqlalchemy>=2.0.36
pydantic>=2.10.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
sentence-transformers>=2.2.0
pgvector>=0.2.0
numpy>=1.26.0
